--- app/src/main/java/com/insurance/chatapp/ChatActivity.kt
@@ -1,28 +1,0 @@
-package com.insurance.chatapp
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatDelegate
-import com.insurance.chatapp.databinding.ActivityChatBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-
-@AndroidEntryPoint
-class ChatActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityChatBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityChatBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.dayNightSwitch?.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-            } else {
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-            }
-        }
-    }
-}
