--- app/src/main/java/com/insurance/chatapp/ChatApp.kt
@@ -1,8 +1,0 @@
-package com.insurance.chatapp
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class ChatApp : Application() {
-}
