--- app/src/main/java/com/insurance/chatapp/common/MainDiffUtil.kt
@@ -1,16 +1,0 @@
-package com.insurance.chatapp.common
-
-import android.annotation.SuppressLint
-import androidx.recyclerview.widget.DiffUtil
-
-class MainDiffUtil<T : Any> : DiffUtil.ItemCallback<T>() {
-
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem === newItem
-    }
-
-    @SuppressLint("DiffUtilEquals")
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem == newItem
-    }
-}
