--- app/src/main/java/com/insurance/chatapp/common/enums/MessageAuthor.kt
@@ -1,6 +1,0 @@
-package com.insurance.chatapp.common.enums
-
-enum class MessageAuthor {
-    TOP,
-    BOTTOM
-}
