--- app/src/main/java/com/insurance/chatapp/domain/models/message/MessageModel.kt
@@ -1,9 +1,0 @@
-package com.insurance.chatapp.domain.models.message
-
-import com.insurance.chatapp.common.enums.MessageAuthor
-
-data class MessageModel(
-    val messageText: String?,
-    val messageAuthor: MessageAuthor,
-    val messageDate: String?
-)
