--- app/src/main/java/com/insurance/chatapp/ui/chat/model/MessageUiModel.kt
@@ -1,6 +1,0 @@
-package com.insurance.chatapp.ui.chat.model
-
-data class MessageUiModel(
-    val messageText: String,
-    val messageDate: String?,
-)
