+++ app/src/main/java/com/space/chatapp/ChatApp.kt
@@ -1,0 +1,8 @@
+package com.space.chatapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class ChatApp : Application() {
+}
