+++ app/src/main/java/com/space/chatapp/common/MainDiffUtil.kt
@@ -1,0 +1,16 @@
+package com.space.chatapp.common
+
+import android.annotation.SuppressLint
+import androidx.recyclerview.widget.DiffUtil
+
+class MainDiffUtil<T : Any> : DiffUtil.ItemCallback<T>() {
+
+    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem === newItem
+    }
+
+    @SuppressLint("DiffUtilEquals")
+    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem == newItem
+    }
+}
