+++ app/src/main/java/com/space/chatapp/common/enums/MessageAuthor.kt
@@ -1,0 +1,6 @@
+package com.space.chatapp.common.enums
+
+enum class MessageAuthor {
+    TOP,
+    BOTTOM
+}
