+++ app/src/main/java/com/space/chatapp/di/ChatMessagesModule.kt
@@ -1,0 +1,19 @@
+package com.space.chatapp.di
+
+import com.space.chatapp.data.local.database.repoitorys.MessageRepositoryImpl
+import com.space.chatapp.domain.repositorys.MessageRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class ChatMessagesModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindMessagesRepository(messageRepositoryImpl: MessageRepositoryImpl): MessageRepository
+
+}
