+++ app/src/main/java/com/space/chatapp/domain/models/message/MessageModel.kt
@@ -1,0 +1,9 @@
+package com.space.chatapp.domain.models.message
+
+import com.space.chatapp.common.enums.MessageAuthor
+
+data class MessageModel(
+    val messageText: String?,
+    val messageAuthor: MessageAuthor,
+    val messageDate: String?
+)
