+++ app/src/main/java/com/space/chatapp/ui/chat/model/MessageUiModel.kt
@@ -1,0 +1,6 @@
+package com.space.chatapp.ui.chat.model
+
+data class MessageUiModel(
+    val messageText: String,
+    val messageDate: String?,
+)
