+++ app/src/main/res/layout/activity_chat.xml
@@ -20,7 +20,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/top_user_fragment"
-        android:name="com.insurance.chatapp.ui.chat.TopUserFragment"
+        android:name="com.space.chatapp.ui.chat.TopUserFragment"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dimen_0_dp"
         android:layout_weight="1"
@@ -34,7 +34,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/bottom_user_fragment"
-        android:name="com.insurance.chatapp.ui.chat.BottomUserFragment"
+        android:name="com.space.chatapp.ui.chat.BottomUserFragment"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dimen_0_dp"
         android:layout_weight="1"
